move time sources for testing into separate files
