+++ RateLimiterTests/TestRateLimiter.cs
@@ -94,39 +94,4 @@ namespace RateLimiterTests
             Assert.AreEqual(RateLimiterConstants.WaitImpossible, rateLimiter.WaitTimeForAmount(1e9));
         }
     }
-
-
-    class TimeSourceZero : ITimeSource
-    {
-        public long CurrentTimeTicks()
-        {
-            return 0L;
-        }
-    }
-    
-    class TimeSourceAdjustable : ITimeSource
-    {
-
-        public TimeSourceAdjustable(long startTime = 0L)
-        {
-            _currentTime = startTime;
-        }
-
-        public long CurrentTimeTicks()
-        {
-            return _currentTime;
-        }
-
-        public void AddTimeTicks(long ticks)
-        {
-            _currentTime += ticks;
-        }
-
-        public void SetTime(long ticks)
-        {
-            _currentTime = ticks;
-        }
-        
-        private long _currentTime;
-    }
 }
