+++ RateLimiterTests/TimeSourceZero.cs
@@ -1,0 +1,12 @@
+using RateLimiter;
+
+namespace RateLimiterTests
+{
+    class TimeSourceZero : ITimeSource
+    {
+        public long CurrentTimeTicks()
+        {
+            return 0L;
+        }
+    }
+}
